Gun pickup, a bit of camera, movement and rotation tweaks
Changed the code to be able to pick up guns, changed raycast to only work on weapons and made it in front of the player kind of. Working on the camera changes such shoulder view so we can actually see something. Made the player only rotate based off of where they look rather than what movement key they press