A lot of things have happened
I rebuilt the collider and functions on enemies to be more consistent and to account for headshots. Btw made headshots. Reworked the perk scripts to make the reset functions easier to use and called less. Made a reset everything function in player controller for when the player goes down. Made most of quick revive and a downed state when you have it. Its janky but it works so good enough for now. Did the outline for the machine. Adjusted auge settings because there was a reloading bug. Made a ui panel for when the player is down so its very obvious. And yeah, probably more stuff. I apparenty spent 6 hours doing a lot of things not related to quick revive