Reworked movement and camera again
Reworked the movement a bit to be based off of rb.MovePosition(targetPosition) so I don't have to manually move the character which was causing jitteriness. The camera was changed completely to be a different camera being ThirdPersonFollow and the Pan Tilt comp to handle rotations. A reason for using this was to use the collision feature so the camera now mostly collides with walls favourably. The only issue I think is in some locations, the camera still clips through the wall due to the player slightly clipping through a wall at certain angles. This could probably be fixed with changing from mesh colliders to something like box colliders but I don't want to deal with it as it works fine now.
Moral of the commit: the movement and camera are as smooth as butter now