Adjusted spread and added a mystery box
Added mystery box that currently has  a 1/16 chance of any gun. Haven't decided on chances of guns yet. It rerolls automatically if it is a dupe so you cannot get a duplicate weapon. Additionally, it deletes the gun if it is not taken after 5 seconds. It doesn't copy the gun, it gives it to you. Additionally, you cannot use it while it has a gun displayed. Adjusted values and function for spread to balance guns a bit and it was only doing vertical spread